Tweaking values
Tweaked various values to do with chunk sizes and simulation size.

Added some debug visualization (press Alt to see)

Fixed bug to do with particles not being set to stable (possibly been around since chunk system was implemented)

New bug appeared after fixing particles not being set to stable: particles will not set other particles to not stable if the other particle is on the other side of a chunk border. This is caused by the loop that unstables neighbouring particles not doing outside the bounds of the current chunk. Removing bounds checking in this loop causes bad errors.